Added Window to show the 2D map of the most recent position of devices

+++ PDS_Detection_System/PDS_Detection_System/CoverageArea.cpp
@@ -50,8 +50,8 @@ bool CoverageArea::onSegment(Point p, Point q, Point r)
 // 2 --> Counterclockwise 
 int CoverageArea::orientation(Point p, Point q, Point r)
 {
-	int val = (q.y - p.y) * (r.x - q.x) -
-		(q.x - p.x) * (r.y - q.y);
+	int val = (int) ( (q.y - p.y) * (r.x - q.x) -
+		(q.x - p.x) * (r.y - q.y) );
 
 	if (val == 0) return 0;  // colinear 
 	return (val > 0) ? 1 : 2; // clock or counterclock wise
